# docs/api/openapi.json
{
  "components": {
    "parameters": {
      "AccountID": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "账号 UUID",
        "schema": {
          "type": "string"
        }
      },
      "Bucket": {
        "name": "bucket",
        "in": "query",
        "description": "桶名；账号有默认桶时可省略",
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "schemas": {
...
      "Error": {
        "type": "object",
        "example": {
          "error": "account not found"
        },
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "S3C_TOKEN 中的 Bearer token。可通过 S3C_TOKEN_SCOPES 为单个 token 限定 readonly / prefixes（桶或桶内键前缀）/ accounts / expiresAt；越权请求返回 403，过期 token 返回 401。未在 S3C_TOKEN_SCOPES 中登记的 token 为全权。",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "s3client API",
    "version": "v1.0.0",
    "description": "s3client 是面向 S3 兼容对象存储的多账号 Web 控制台。本文档为 /api/* 端点的 OpenAPI 3.0 契约，所有响应均 JSON（除 /api/health 等纯状态端点）。鉴权：Bearer Token（环境变量 S3C_TOKEN，多值逗号分隔）。"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/accounts/{id}/bucket/object-lock": {
      "get": {
        "operationId": "getObjectLock",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          },
          {
            "$ref": "#/components/parameters/Bucket"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "bucket": "my-bucket",
                  "enabled": true,
                  "defaultRetentionMode": "GOVERNANCE",
                  "defaultRetentionDays": 30,
                  "defaultRetentionYears": 0
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string"
                    },
                    "defaultRetentionDays": {
                      "type": "integer"
                    },
                    "defaultRetentionMode": {
                      "type": "string",
                      "description": "GOVERNANCE | COMPLIANCE（未配置默认保留时为空串）"
                    },
                    "defaultRetentionYears": {
                      "type": "integer"
                    },
                    "enabled": {
                      "type": "boolean"
                    }
                  }
                }
              }
            },
            "description": "未启用 Object Lock 时 enabled=false"
          }
        },
        "summary": "桶 Object Lock 配置",
        "tags": [
          "bucket-settings"
        ]
      },
      "put": {
        "operationId": "putObjectLock",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "defaultRetentionMode": "GOVERNANCE",
                "defaultRetentionDays": 30,
                "defaultRetentionYears": 0
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "defaultRetentionDays": {
                    "type": "integer",
                    "description": "与 defaultRetentionYears 二选一；必须 ≥1"
                  },
                  "defaultRetentionMode": {
                    "type": "string",
                    "enum": [
                      "GOVERNANCE",
                      "COMPLIANCE"
                    ]
                  },
                  "defaultRetentionYears": {
                    "type": "integer",
                    "description": "与 defaultRetentionDays 二选一；必须 ≥1"
                  }
                },
                "required": [
                  "bucket",
                  "defaultRetentionMode"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "bucket": "my-bucket",
                  "enabled": true,
                  "defaultRetentionMode": "GOVERNANCE",
                  "defaultRetentionDays": 30,
                  "defaultRetentionYears": 0
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string"
                    },
                    "defaultRetentionDays": {
                      "type": "integer"
                    },
                    "defaultRetentionMode": {
                      "type": "string",
                      "description": "GOVERNANCE | COMPLIANCE（未配置默认保留时为空串）"
                    },
                    "defaultRetentionYears": {
                      "type": "integer"
                    },
                    "enabled": {
                      "type": "boolean"
                    }
                  }
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "输入非法 / 桶未启用 Object Lock"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "桶未在创建时启用 Object Lock（InvalidBucketState）"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "厂商未实现 Object Lock（NotImplemented）"
          }
        },
        "summary": "设置桶默认保留策略（桶须创建时启用 Object Lock）",
        "tags": [
          "bucket-settings"
        ]
      }
    },
...
    "/api/accounts/{id}/object-legal-hold": {
      "get": {
        "operationId": "getObjectLegalHold",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          },
          {
            "$ref": "#/components/parameters/Bucket"
          },
          {
            "description": "",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "可选；读取指定版本的法定保留",
            "in": "query",
            "name": "versionId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "bucket": "my-bucket",
                  "key": "docs/a.txt",
                  "versionId": "",
                  "status": "ON"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "versionId": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "未设置 → status=OFF"
          }
        },
        "summary": "对象法定保留状态",
        "tags": [
          "object-meta"
        ]
      },
      "put": {
        "operationId": "putObjectLegalHold",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "key": "docs/a.txt",
                "versionId": "",
                "status": "ON"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "key": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ON",
                      "OFF"
                    ]
                  },
                  "versionId": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "key",
                  "status"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "bucket": "my-bucket",
                  "key": "docs/a.txt",
                  "versionId": "",
                  "status": "ON"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "versionId": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "OK"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "拒绝（含越权）"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "对象被锁定"
          }
        },
        "summary": "设置对象法定保留（ON/OFF）",
        "tags": [
          "object-meta"
        ]
      }
    },
    "/api/accounts/{id}/object-retention": {
      "get": {
        "operationId": "getObjectRetention",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          },
          {
            "$ref": "#/components/parameters/Bucket"
          },
          {
            "description": "",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "可选；读取指定版本的保留期",
            "in": "query",
            "name": "versionId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "bucket": "my-bucket",
                  "key": "docs/a.txt",
                  "versionId": "",
                  "configured": true,
                  "mode": "COMPLIANCE",
                  "retainUntilDate": "2030-01-02T03:04:05Z"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string"
                    },
                    "configured": {
                      "type": "boolean"
                    },
                    "key": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "description": "GOVERNANCE | COMPLIANCE（configured=false 时为空串）"
                    },
                    "retainUntilDate": {
                      "type": "string",
                      "description": "RFC3339 到期时间（configured=false 时为空串）"
                    },
                    "versionId": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "无保留期（或桶未启用 Object Lock）时 configured=false"
          }
        },
        "summary": "对象保留期（Object Lock）",
        "tags": [
          "object-meta"
        ]
      },
      "put": {
        "operationId": "putObjectRetention",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "key": "docs/a.txt",
                "versionId": "",
                "mode": "GOVERNANCE",
                "retainUntilDate": "2031-02-03T04:05:06Z"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "key": {
                    "type": "string"
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "GOVERNANCE",
                      "COMPLIANCE"
                    ]
                  },
                  "retainUntilDate": {
                    "type": "string",
                    "description": "RFC3339（如 2031-02-03T04:05:06Z），必须是未来时刻"
                  },
                  "versionId": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "key",
                  "mode",
                  "retainUntilDate"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "bucket": "my-bucket",
                  "key": "docs/a.txt",
                  "versionId": "",
                  "configured": true,
                  "mode": "GOVERNANCE",
                  "retainUntilDate": "2031-02-03T04:05:06Z"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string"
                    },
                    "configured": {
                      "type": "boolean"
                    },
                    "key": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string"
                    },
                    "retainUntilDate": {
                      "type": "string"
                    },
                    "versionId": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "输入非法 / 保留期违规"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "GOVERNANCE 保留期内的拒绝（含越权）"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "对象被 COMPLIANCE 锁定（ObjectLocked）"
          }
        },
        "summary": "设置对象保留期（Object Lock）",
        "tags": [
          "object-meta"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/",
      "description": "同源（前端 Vite 代理或后端 SPA fallback）"
    }
  ],
...
}